backend1/requirements.txt
flask_cors
flask_sqlalchemy
flask_session